+++ app/app.go
@@ -121,10 +121,16 @@ import (
 	"github.com/spf13/cast"
 
 	"github.com/archway-network/archway/wasmbinding"
+
+	"github.com/archway-network/archway/x/callback"
+	callbackKeeper "github.com/archway-network/archway/x/callback/keeper"
+	callbackTypes "github.com/archway-network/archway/x/callback/types"
+
 	"github.com/archway-network/archway/x/rewards"
 	rewardsKeeper "github.com/archway-network/archway/x/rewards/keeper"
 	"github.com/archway-network/archway/x/rewards/mintbankkeeper"
 	rewardsTypes "github.com/archway-network/archway/x/rewards/types"
+
 	"github.com/archway-network/archway/x/tracking"
 	trackingKeeper "github.com/archway-network/archway/x/tracking/keeper"
 	trackingTypes "github.com/archway-network/archway/x/tracking/types"
@@ -200,6 +206,7 @@ var (
 		tracking.AppModuleBasic{},
 		rewards.AppModuleBasic{},
 		genmsg.AppModule{},
+		callback.AppModuleBasic{},
 	)
 
 	// module account permissions
@@ -217,6 +224,7 @@ var (
 		icatypes.ModuleName:                  nil,
 		wasmdTypes.ModuleName:                {authtypes.Burner},
 		rewardsTypes.TreasuryCollector:       {authtypes.Burner},
+		callbackTypes.ModuleName:             nil,
 	}
 )
 
@@ -296,7 +304,7 @@ func NewArchwayApp(
 		feegrant.StoreKey, authzkeeper.StoreKey, wasmdTypes.StoreKey, consensusparamtypes.StoreKey,
 		icahosttypes.StoreKey, ibcfeetypes.StoreKey, crisistypes.StoreKey, group.StoreKey, nftkeeper.StoreKey,
 
-		trackingTypes.StoreKey, rewardsTypes.StoreKey,
+		trackingTypes.StoreKey, rewardsTypes.StoreKey, callbackTypes.StoreKey,
 	)
 	tkeys := sdk.NewTransientStoreKeys(paramstypes.TStoreKey)
 	memKeys := sdk.NewMemoryStoreKeys(capabilitytypes.MemStoreKey)
@@ -555,6 +563,15 @@ func NewArchwayApp(
 		govModuleAddr,
 	)
 
+	app.Keepers.CallbackKeeper = callbackKeeper.NewKeeper(
+		appCodec,
+		keys[callbackTypes.StoreKey],
+		app.Keepers.WASMKeeper,
+		app.Keepers.RewardsKeeper,
+		app.Keepers.BankKeeper,
+		govModuleAddr,
+	)
+
 	var transferStack porttypes.IBCModule
 	transferStack = transfer.NewIBCModule(app.Keepers.TransferKeeper)
 	transferStack = ibcfee.NewIBCMiddleware(transferStack, app.Keepers.IBCFeeKeeper)
@@ -628,6 +645,7 @@ func NewArchwayApp(
 		tracking.NewAppModule(app.appCodec, app.Keepers.TrackingKeeper),
 		rewards.NewAppModule(app.appCodec, app.Keepers.RewardsKeeper),
 		genmsg.NewAppModule(app.MsgServiceRouter()),
+		callback.NewAppModule(app.appCodec, app.Keepers.CallbackKeeper, app.Keepers.WASMKeeper),
 		crisis.NewAppModule(&app.Keepers.CrisisKeeper, skipGenesisInvariants, app.getSubspace(crisistypes.ModuleName)), // always be last to make sure that it checks for all invariants and not only part of them
 	)
 
@@ -666,6 +684,7 @@ func NewArchwayApp(
 		// wasm gas tracking
 		trackingTypes.ModuleName,
 		rewardsTypes.ModuleName,
+		callbackTypes.ModuleName,
 	)
 
 	app.mm.SetOrderEndBlockers(
@@ -698,6 +717,7 @@ func NewArchwayApp(
 		// wasm gas tracking
 		trackingTypes.ModuleName,
 		rewardsTypes.ModuleName,
+		callbackTypes.ModuleName,
 		// invariants checks are always the last to run
 		crisistypes.ModuleName,
 	)
@@ -739,6 +759,7 @@ func NewArchwayApp(
 		// wasm gas tracking
 		trackingTypes.ModuleName,
 		genmsg.ModuleName,
+		callbackTypes.ModuleName,
 		// invariants checks are always the last to run
 		crisistypes.ModuleName,
 	)