+++ app/keepers/keepers.go
@@ -25,6 +25,7 @@ import (
 	ibctransferkeeper "github.com/cosmos/ibc-go/v7/modules/apps/transfer/keeper"
 	ibckeeper "github.com/cosmos/ibc-go/v7/modules/core/keeper"
 
+	callbackKeeper "github.com/archway-network/archway/x/callback/keeper"
 	rewardsKeeper "github.com/archway-network/archway/x/rewards/keeper"
 	trackingKeeper "github.com/archway-network/archway/x/tracking/keeper"
 )
@@ -52,6 +53,9 @@ type ArchwayKeepers struct {
 	FeeGrantKeeper        feegrantkeeper.Keeper
 	AuthzKeeper           authzkeeper.Keeper
 	WASMKeeper            wasmkeeper.Keeper
-	TrackingKeeper        trackingKeeper.Keeper
-	RewardsKeeper         rewardsKeeper.Keeper
+
+	// Archway Keepers
+	TrackingKeeper trackingKeeper.Keeper
+	RewardsKeeper  rewardsKeeper.Keeper
+	CallbackKeeper callbackKeeper.Keeper
 }