+++ contracts/callback-test/src/msg.rs
@@ -1,0 +1,30 @@
+use cosmwasm_schema::{cw_serde, QueryResponses};
+
+#[cw_serde]
+pub struct InstantiateMsg {
+    pub count: i32,
+}
+
+#[cw_serde]
+pub enum ExecuteMsg {
+    Increment {},
+    Reset { count: i32 },
+}
+
+#[cw_serde]
+pub enum SudoMsg {
+    Callback { job_id: u64 },
+}
+
+#[cw_serde]
+#[derive(QueryResponses)]
+pub enum QueryMsg {
+    // GetCount returns the current count as a json-encoded number
+    #[returns(GetCountResponse)]
+    GetCount {},
+}
+
+#[cw_serde]
+pub struct GetCountResponse {
+    pub count: i32,
+}